if natural-scroll is enabled, invert scroll inputs
When natural scrolling setting is enabled, the scroll inputs are essentially the
opposite of what's actually physically happening. This is fine for scrolling
down a page, but feels very weird for things like volume control.

Invert natural scrolling inputs so scrolling up actually increases the volume,
and vice versa.

Fixes #112.