Fix label fitering cutting at the first hyphen
The label filtering will now only remove the text
containing a filtered word from the last "separator" encountered
and not the first.

Also now the spaces before that separator are removed too,
so there are no trailing spaces in the end result.

(The separator being a hyphen or an opening bracket.)